Split subscribers into a separate infrastructure:subscribers module.
