+++ infrastructure/events/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   implementation project(':data:metrics')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:subscribers')
   implementation project(':util')
 
   testImplementation 'org.hyperledger.besu.internal:metrics-core'
