+++ infrastructure/events/src/main/java/tech/pegasys/teku/infrastructure/events/EventDeliverer.java
@@ -21,8 +21,8 @@ import org.hyperledger.besu.plugin.services.metrics.Counter;
 import org.hyperledger.besu.plugin.services.metrics.LabelledMetric;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.subscribers.Subscribers;
 import tech.pegasys.teku.metrics.TekuMetricCategory;
-import tech.pegasys.teku.util.events.Subscribers;
 
 abstract class EventDeliverer<T> {
   private final Subscribers<T> subscribers = Subscribers.create(true);
