+++ infrastructure/subscribers/src/test/java/tech/pegasys/teku/infrastructure/subscribers/SubscribersTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.events;
+package tech.pegasys.teku.infrastructure.subscribers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
