+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/methods/MetadataMessagesFactory.java
@@ -16,10 +16,10 @@ package tech.pegasys.teku.networking.eth2.rpc.beaconchain.methods;
 import java.util.concurrent.atomic.AtomicLong;
 import tech.pegasys.teku.datastructures.networking.libp2p.rpc.MetadataMessage;
 import tech.pegasys.teku.datastructures.networking.libp2p.rpc.PingMessage;
+import tech.pegasys.teku.infrastructure.subscribers.ValueObserver;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.ssz.SSZTypes.Bitvector;
 import tech.pegasys.teku.util.config.Constants;
-import tech.pegasys.teku.util.events.ValueObserver;
 
 public class MetadataMessagesFactory implements ValueObserver<Iterable<Integer>> {
