+++ pow/build.gradle
@@ -4,9 +4,10 @@ dependencies {
   implementation project(':ssz')
   implementation project(':util')
   implementation project(':infrastructure:events')
+  implementation project(':infrastructure:logging')
+  implementation project(':infrastructure:subscribers')
   implementation project(':services:serviceutils')
   implementation project(':storage:api')
-  implementation project(':infrastructure:logging')
 
   api 'org.web3j:core'
