+++ settings.gradle
@@ -20,6 +20,7 @@ include 'infrastructure:events'
 include 'infrastructure:io'
 include 'infrastructure:unsigned'
 include 'infrastructure:logging'
+include 'infrastructure:subscribers'
 include 'networking:eth2'
 include 'networking:p2p'
 include 'services'
